scripts/requirements_gui.txt
pandas>=1.5.0
openpyxl>=3.0.0
tkinter
pillow>=9.0.0 